40. Добавление AdminPolicy, PostPolicy. Редактирование post.show

+++ app/Http/Controllers/Personal/Post/DeleteController.php
@@ -4,12 +4,14 @@ namespace App\Http\Controllers\Personal\Post;
 
 use App\Http\Controllers\Controller;
 use App\Models\Post;
+use Illuminate\Support\Facades\Gate;
 
 
 class DeleteController extends Controller
 {
     public function __invoke(Post $post)
     {
+        Gate::authorize('delete', $post);
         $post->delete();
         return redirect()->route('personal.post.index');
     }
